(#72) - Refactored boilerplate user/log creation to it's own module
